Clean up TestBuilder OverlayNG functions
